aas-sm-to-py/pyproject.toml
[build-system]
requires = ["setuptools>=61", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "sm-generator"
version = "0.1.1"
description = "Generate Submodel-Specific classes with filled metainformation derived from submodel templates"
requires-python = ">=3.7"
license = { text = "MIT" }
authors = [
  { name = "Igor Garmaev", email = "garmaev@gmx.net" }
]
keywords = ["aas", "submodel", "class", "code", "generation", "python"]
classifiers = [
  "Intended Audience :: Developers",
  "Programming Language :: Python :: 3",
  "License :: OSI Approved :: MIT License",
  "Operating System :: OS Independent"
]
dependencies = [
  "basyx-python-sdk @ git+https://github.com/eclipse-basyx/basyx-python-sdk.git@develop#subdirectory=sdk",
  "jinja2>=3.1.2,<4",
  "black"
]

[project.urls]
Homepage = "https://github.com/zrgt/aas-submodel-template-to-py"

[project.scripts]
submodel_to_code = "asttp.submodel_to_code:main"

[tool.setuptools.packages.find]
where = ["."]

[tool.setuptools.package-data]
asttp = ["code_templates/*.pyi"]